finalizado!!!""

+++ src/main/java/com/sopra/springboot/app/models/service/UploadFileServiceImpl.java
@@ -27,9 +27,9 @@ public class UploadFileServiceImpl implements IUploadFileService {
 	public Resource load(String filename) throws MalformedURLException {
 		Path pathFoto = getPath(filename);
 		log.info("pathFoto: " + pathFoto);
-		
+
 		Resource recurso = new UrlResource(pathFoto.toUri());
-		
+
 		if (!recurso.exists() || !recurso.isReadable()) {
 			throw new RuntimeException("Error: no se puede cargar la imagen: " + pathFoto.toString());
 		}
@@ -50,12 +50,14 @@ public class UploadFileServiceImpl implements IUploadFileService {
 
 	@Override
 	public boolean delete(String filename) {
-		Path rootPath = getPath(filename);
-		File archivo = rootPath.toFile();
-		
-		if(archivo.exists() && archivo.canRead()) {
-			if(archivo.delete()) {
-				return true;
+		if (filename != null && filename.length() > 0) {
+			Path rootPath = getPath(filename);
+			File archivo = rootPath.toFile();
+
+			if (archivo.exists() && archivo.canRead()) {
+				if (archivo.delete()) {
+					return true;
+				}
 			}
 		}
 		return false;
@@ -68,7 +70,7 @@ public class UploadFileServiceImpl implements IUploadFileService {
 	@Override
 	public void deleteAll() {
 		FileSystemUtils.deleteRecursively(Paths.get(UPLOADS_FOLDER).toFile());
-		
+
 	}
 
 	@Override
